Add Mongolian product context to market-info endpoint
Products with context:
- MN-COAL-PERP: Tavan Tolgoi benchmark, FOB Gashuun Sukhait
- MN-CASHMERE-PERP: White dehairing, 14.5-16.0 microns
- MN-COPPER-PERP: Oyu Tolgoi concentrate, LME Grade A
- MN-MEAT-PERP: Mutton composite, Ulaanbaatar delivery
- MN-REALESTATE-PERP: UB apartments, per square meter

+++ src/http_server.h
@@ -1104,7 +1104,7 @@ inline void HttpServer::setup_routes() {
         response["mark_price_mnt"] = product->mark_price_mnt;
 
         // Add Mongolian product context
-        if (product->category == static_cast<int>(ProductCategory::MN_PERPETUAL) || 
+        if (product->category == ProductCategory::MN_PERPETUAL || 
             symbol.find("MN-") == 0) {
             
             json context;
